refactor: rename all hooks and components to kebab-case
Files renamed:
- useCriticalAction.ts → use-critical-action.ts
- useLocalCache.ts → use-local-cache.ts
- useServerTick.ts → use-server-tick.ts
- useLens.ts → use-lens.ts
- useABToggle.ts → use-ab-toggle.ts
- SigilZone.tsx → sigil-zone.tsx
- CriticalZone.tsx → critical-zone.tsx
- MachineryLayout.tsx → machinery-layout.tsx
- GlassLayout.tsx → glass-layout.tsx
- LensProvider.tsx → lens-provider.tsx
- CriticalButton.tsx → critical-button.tsx
- GlassButton.tsx → glass-button.tsx
- MachineryItem.tsx → machinery-item.tsx
- Button.tsx → button.tsx
- All recipe files (ConfirmFlow, HeroCard, etc.)

All imports updated to match new file names.
