+++ sigil-mark/components/index.ts
@@ -24,4 +24,4 @@
  * ```
  */
 
-export { Button, type ButtonProps } from './Button';
+export { Button, type ButtonProps } from './button';
