+++ sigil-mark/layouts/machinery-layout.tsx
