+++ sigil-mark/lenses/default/machinery-item.tsx
