+++ sigil-mark/lenses/strict/machinery-item.tsx
